+++ apps/dashboard/app/javascript/utils.js
@@ -1,3 +1,4 @@
+import {analyticsPath} from "./config";
 
 export function cssBadgeForState(state){
   switch (state) {
@@ -60,11 +61,17 @@ export function bindFullPageSpinnerEvent() {
 // open links in javascript and display an alert
 export function openLinkInJs(event) {
   event.preventDefault();
-  const href = event.target.href;
+  let href = event.target.href;
 
-  // do nothing if there's no href.
-  if(href == null){
-    return;
+  // event.target could be a child of the anchor, so try that.
+  if(href == null) {
+    const closestAnchor = event.target.closest('a');
+    if(closestAnchor.hasChildNodes(event.target)) {
+      href = closestAnchor.href;
+    } else {
+      // event.target is not a child of an anhcor, so there's nothing to do.
+      return;
+    }
   }
 
   if(window.open(href) == null) {
@@ -101,4 +108,13 @@ export function setInnerHTML(element, html) {
 
     currentElement.parentNode.replaceChild(newElement, currentElement);
   });
+}
+
+// Helper method to report errors from the front end via AJAX
+export function reportErrorForAnalytics(path, error) {
+  // error - report back for analytics purposes
+  const analyticsUrl = new URL(analyticsPath(path), document.location);
+  analyticsUrl.searchParams.append('error', error);
+  // Fire and Forget
+  fetch(analyticsUrl);
 }