+++ apps/dashboard/app/lib/smart_attributes.rb
@@ -10,6 +10,7 @@ module SmartAttributes
   require 'smart_attributes/attributes/auto_groups'
   require 'smart_attributes/attributes/auto_job_name'
   require 'smart_attributes/attributes/auto_modules'
+  require 'smart_attributes/attributes/auto_log_location'
   require 'smart_attributes/attributes/auto_primary_group'
   require 'smart_attributes/attributes/auto_queues'
   require 'smart_attributes/attributes/auto_qos'
@@ -23,4 +24,5 @@ module SmartAttributes
   require 'smart_attributes/attributes/bc_vnc_resolution'
   require 'smart_attributes/attributes/auto_environment_variable'
   require 'smart_attributes/attributes/auto_cores'
+  require 'smart_attributes/attributes/global_attribute'
 end