+++ apps/dashboard/test/models/motd_formatter/motd_formatter_plaintext_test.rb
@@ -1,9 +1,10 @@
+# frozen_string_literal: true
+
 require 'test_helper'
 
 class MotdFormatterPlaintextTest < ActiveSupport::TestCase
   include MotdFormatter
-  test "test when motd formatter_plaintext_valid" do
-
+  test 'test when motd formatter_plaintext_valid' do
     path = "#{Rails.root}/test/fixtures/files/motd_valid"
     motd_file = MotdFile.new(path)
     formatted_motd = MotdFormatterPlaintext.new(motd_file)
@@ -12,7 +13,7 @@ class MotdFormatterPlaintextTest < ActiveSupport::TestCase
     assert_equal expected_file, formatted_motd.content
   end
 
-  test "test when motd_formatter_plaintext empty" do
+  test 'test when motd_formatter_plaintext empty' do
     path = "#{Rails.root}/test/fixtures/files/motd_empty"
     motd_file = MotdFile.new(path)
     formatted_motd = MotdFormatterPlaintext.new(motd_file)
@@ -20,7 +21,7 @@ class MotdFormatterPlaintextTest < ActiveSupport::TestCase
     assert_equal '', formatted_motd.content
   end
 
-  test "test when motd formatter_plaintext_missing" do
+  test 'test when motd formatter_plaintext_missing' do
     path = "#{Rails.root}/test/fixtures/files/motd_missing"
     motd_file = MotdFile.new(path)
     formatted_motd = MotdFormatterPlaintext.new(motd_file)
@@ -28,7 +29,7 @@ class MotdFormatterPlaintextTest < ActiveSupport::TestCase
     assert_equal '', formatted_motd.content
   end
 
-  test "test when motd formatter_plaintext_nil" do
+  test 'test when motd formatter_plaintext_nil' do
     motd_file = nil
     formatted_motd = MotdFormatterPlaintext.new(motd_file)